backend/requirements.txt
fastapi
uvicorn
pyswip
psycopg2-binary
pydantic-settings
